Moved service file into v1 folder. Added export to index file for service file. Updated import int service.test
