+++ lib/api/patch/v1/index.ts
@@ -1,1 +1,2 @@
 export * from "./types";
+export * from "./service";
